Goes back to the 'material' theme
It's a little too much trouble to either 'cinder' or 'mkdocs' work
nicely with all the extra little features that are already present in
'material'.

So while the 'material' theme looks a bit too complicated and targeted
for developers, it frees me up to thinking about just the documentation
itself.

Will stick with 'material' for now.